Most UIs now work with a gamepad aswell as mouse and keyboard
You can use the select button to open the journal, shoulder buttons to switch between the tabs and each screen should allow you to use the information within it. Dialogue boxes now work with gamepad, aswell as quest accept and complete interfaces.

The player can now interact with the game using a gamepad