Fixed Ranged
An arrow will now arc correctly, and will fly towards the center of the screen. It's gravity may need to be altered at somepoint.

The projectile speed changes depending on how long the mouse button has been held down
Fixed an issue with any chest inventories not working with the controller
Allowed items to now be used with a controller
Player's animation will reset normally when exiting a lake