We have roaming goats and bears!!!! And Waterfall
Bear anims are shite, we cant have sleep or eat however Goats will eat after a duration and run away soon! I have added the waterfall and also the audio for it but need to do it by distance still. I have also updated the shrine with nice rocks and I have made a large bear and a small bear in the shrine to represent blair and artio!